combine events and testimonials in backend
