+++ backend/src/app.ts
@@ -9,7 +9,6 @@ import { isHttpError } from "http-errors";
 import subscriberRoutes from "src/routes/subscriber";
 import memberRoutes from "src/routes/members";
 import testimonialRoutes from "src/routes/testimonial";
-import eventRoutes from "src/routes/events";
 
 const app = express();
 
@@ -31,7 +30,6 @@ app.use(
 app.use("/api/subscribers", subscriberRoutes);
 app.use("/api/member", memberRoutes);
 app.use("/api/testimonial", testimonialRoutes);
-app.use("/api/events", eventRoutes);
 /**
  * Error handler; all errors thrown by server are handled here.
  * Explicit typings required here because TypeScript cannot infer the argument types.
