+++ backend/src/controllers/testimonial.ts
@@ -2,13 +2,16 @@ import { RequestHandler } from "express";
 import TestimonialModel from "src/models/testimonial";
 
 export const createTestimonial: RequestHandler = async (req, res, next) => {
-  const { quote, description, image } = req.body;
+  const { title, description, image, type } = req.body;
+  console.log("here");
   try {
     const testimonial = await TestimonialModel.create({
-      quote: quote,
+      title: title,
       description: description,
       image: image,
+      type: type,
     });
+    console.log("here");
     res.status(201).json(testimonial);
   } catch (error) {
     next(error);
