+++ backend/src/models/testimonial.ts
@@ -1,9 +1,10 @@
 import { InferSchemaType, Schema, model } from "mongoose";
 
 const testimonialSchema = new Schema({
-  quote: { type: String, required: true },
+  title: { type: String, required: true },
   description: { type: String, required: true },
   image: { type: String, required: true },
+  type: { type: String, required: true },
 });
 
 type testimonial = InferSchemaType<typeof testimonialSchema>;
