+++ frontend/src/components/eventCard.css
