+++ frontend/src/components/testimonialCard.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import React from "react";
 
 import "./testimonialCard.css";
-import type { Testimonial } from "../../api/testimonial";
+import type { Testimonial } from "../api/testimonial";
 
 type CardProps = {
   testimonial: Testimonial | null | undefined;
@@ -14,7 +14,7 @@ const TestimonialCard = ({ testimonial }: CardProps) => {
       <main className="testimonial-card-container">
         <Image src={testimonial.image} alt="image" width={400} height={200} />
         <div className="testimonial-text-container">
-          <h1>{testimonial.quote}</h1>
+          <h1>{testimonial.title}</h1>
           <p>{testimonial.description}</p>
         </div>
       </main>
